+++ lib/models/event.g.dart
@@ -6,13 +6,12 @@ part of 'event.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$EventImpl _$$EventImplFromJson(Map<String, dynamic> json) => _$EventImpl(
+_$_Event _$$_EventFromJson(Map<String, dynamic> json) => _$_Event(
       date: DateTime.parse(json['date'] as String),
       location: $enumDecode(_$LocationEnumMap, json['location']),
     );
 
-Map<String, dynamic> _$$EventImplToJson(_$EventImpl instance) =>
-    <String, dynamic>{
+Map<String, dynamic> _$$_EventToJson(_$_Event instance) => <String, dynamic>{
       'date': instance.date.toIso8601String(),
       'location': _$LocationEnumMap[instance.location]!,
     };